requirements.txt
opencv-python-headless>=4.8.0
ffmpeg-python>=0.2.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
plotly>=5.15.0
scikit-image>=0.20.0